Test/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
apturl==0.5.2
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
cached-property==1.5.1
certifi==2023.5.7
chardet==3.0.4
charset-normalizer==3.2.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dateparser==1.1.8
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info==0.23+ubuntu1.1
Django==4.2.4
django-crispy-forms==2.0
django-mpesa==2.0.14
djangorestframework==3.14.0
dnspython==1.16.0
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.21.7
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
frozenlist==1.3.3
future==0.18.2
greenlet==2.0.2
httplib2==0.14.0
idna==3.4
importlib-metadata==1.5.0
inflection==0.5.1
JayDeBeApi==1.2.3
JPype1==1.4.1
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
more-itertools==4.2.0
multidict==6.0.4
mysql-connector-python==8.1.0
mysqlclient==2.2.0
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
packaging==23.1
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
protobuf==4.21.12
psycopg2==2.9.6
pycairo==1.16.2
pycrypto==2.6.1
pycryptodome==3.18.0
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt==2.0.1+ubuntu0.20.4.1
python-binance==1.0.17
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
python-decouple==3.8
pytz==2023.3
pyxdg==0.26
PyYAML==6.0.1
regex==2023.6.3
reportlab==3.5.34
requests==2.31.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.16.0
SQLAlchemy==2.0.19
sqlanydb==1.0.13
sqlparse==0.4.4
ssh-import-id==5.10
systemd-python==234
texttable==1.6.2
typing-extensions==4.7.1
tzlocal==5.0.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
ujson==5.8.0
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.0.3
usb-creator==0.3.7
wadllib==1.3.3
watchdog==3.0.0
websocket-client==0.53.0
websockets==11.0.3
xkit==0.0.0
yarl==1.9.2
zipp==1.0.0
